api/requirements.txt
Flask
flask
requests
flask-cors
python-dotenv
boto3
google-api-python-client
google-auth